Dealt with size_t to int warnings

+++ TriviaProject/TriviaProject/Communicator.cpp
@@ -175,7 +175,7 @@ In: The data buffer + the socket
 void Communicator::sendData(SOCKET clientSocket, std::vector<unsigned char>& data)
 {
 	const char* temp = reinterpret_cast<const char *>(data.data());
-	if (send(clientSocket, temp, data.size(), 0) == SOCKET_ERROR)
+	if (send(clientSocket, temp, (int)data.size(), 0) == SOCKET_ERROR)
 	{
 		throw std::exception("Error while sending message to client, Socket error - " + WSAGetLastError());
 	}
